requirements.txt
Flask == 2.1.2
Werkzeug~=2.1.2
PyPDF2~=2.4.1
google-cloud-storage==2.1.0
